feat: S3 secret environment variable
Tweak to how S3 secret is read. First the AppSetting for S3SecretKey is read, if that is null then it searches for S3SecretEnvironmentVariable for a tenant level env var.

This will initially not be present in agreements/licenses, and so null/empty-safeing has been introduced to AppSetting.getSettingValue.

Finally if all above have returned null, fall back to GLOBAL_S3_SECRET_KEY