feat: default 500 handling (#17)
* feat: ErrorController for default handling of 500s

Included ErrorController for a default 500 response view across our KIWT modules.

Includes the ability to turn on/off stacktraces in response via config property 'webtoolkit.endpoints.includeStackTraceFor500'

* feat: Add custom exception for SimpleLookupService

Also set up interface `errorHandleable500`, which allows for a custom message on a sanitised exception to be filtered through to the 500 default catch route.

* chore: Extra comments for clarity

* refactor: Exception handling interface no longer assumes end result of 500.

Can spin up "ErrorHandle" with a code and a message, which will be reflected in response

* chore: Tweak name of application config property to webtoolkit.endpoints.includeStackTrace

* chore: Change grails application config to `endpoints.include-stack-trace`, removing namespace webtoolkit

* test: Add unit test for ErrorController

* chore: Actually use INVALID_PROPERTY static, rather than using it implicitly by passing 1

* chore: Removed commented out test code (whoops)